src/Backend/chatbot-backend/requirements.txt
appwrite
openai
flask
flask_cors
python-dotenv